separate the solution into two implementations
