+++ tests/Divide_Conquer/315.Count-of-Smaller-Numbers-After-Self/315.binary-indexed-tree.h
@@ -1,0 +1,12 @@
+#ifndef TESTS_DIVIDE_CONQUER_315_BINARY_INDEXED_TREE
+#define TESTS_DIVIDE_CONQUER_315_BINARY_INDEXED_TREE
+
+#include <vector>
+#include <algorithm>
+using namespace std;
+#include <assert.h>
+
+// binary indexed tree
+extern "C" { std::vector<int> count_v2(std::vector<int> &nums); }
+
+#endif
