+++ tests/Divide_Conquer/315.Count-of-Smaller-Numbers-After-Self/315.divided-conquer.h
@@ -1,0 +1,12 @@
+#ifndef TESTS_DIVIDE_CONQUER_315_DIVIDE_CONQUER
+#define TESTS_DIVIDE_CONQUER_315_DIVIDE_CONQUER
+
+#include <vector>
+#include <algorithm>
+using namespace std;
+#include <assert.h>
+
+// divided conquer
+extern "C" { std::vector<int> count_v1(std::vector<int> &nums); }
+
+#endif
